refactor: create deletion authorization behavior so it can be reused in the future
